Code for generating stars.
Wrote the code to generate a bunch of randomized stars scattered around us in the sky.  By design, a bunch will be out of view beneath the horizon or behind us; part of what I'd like to do is trivially hide the ones that the user can't currently see.

Doesn't work yet -- my numbers are 'way out of whack.